front-end auth

+++ back-end/restful-web-services/src/main/java/com/in28minutes/rest/webservices/restfulwebservices/todo/TodoHardcodedService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class TodoHardcodedService {
 
     private static List<Todo> todos = new ArrayList<>();
-    private static int idCounter = 0;
+    private static Long idCounter = 0L;
     static {
         todos.add(new Todo(++idCounter, "in28Minutes", "learn to dance", new Date(), false));
         todos.add(new Todo(++idCounter, "in28Minutes", "learn about microservices 2", new Date(), false));
@@ -32,7 +32,7 @@ public class TodoHardcodedService {
         return todo;
     }
 
-    public Todo deleteById(long id) {
+    public Todo deleteById(Long id) {
         Todo todo = findById(id);
         if(todo==null) return null;
 
@@ -42,7 +42,7 @@ public class TodoHardcodedService {
         return null;
     }
 
-    public Todo findById(long id) {
+    public Todo findById(Long id) {
         for(Todo todo: todos) {
             if(todo.getId() == id) {
                 return todo;
